# specs/003-seo-optimization/contracts/seo-openapi.yaml
openapi: 3.0.3
info:
  title: Crickzen SEO API
  version: 1.0.0
  description: Endpoints supporting SEO artifacts and diagnostics
servers:
  - url: https://www.crickzen.com
paths:
...
  /sitemaps/{name}.xml:
    get:
      summary: Public sitemap partition (XML)
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          description: Partition filename stem (e.g., sitemap-matches-0001)
      responses:
        '200':
          description: XML sitemap partition (edge handles gzip)
          content:
            application/xml:
              schema:
                type: string
                format: binary
...